+++ NEWS.md
@@ -1,3 +1,14 @@
+# cntYearTrends 0.2.0 (29.05.2023)
+
+## New features
+
+-   Distributions of responses to the items returned as additional element of simulation results.
+
+## Bug fixes
+
+-   `run_simulation()` doesn't fail due to wrongly assessing number of columns storing items' threshold values.
+-   Data generating functions consequently use integer as a format for variables storing project, country, year and item ids.
+
 # cntYearTrends 0.1.0 (28.05.2023)
 
 -   First usable version of the package - see README.md.